pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraud-fl-dp"
version = "1.0.0"
description = "Privacy-preserving federated learning for financial fraud detection (DP-CTGAN + DP-FedAvg)"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Ifeanyi Bryan Uzoatu" },
]
keywords = [
    "federated learning",
    "differential privacy",
    "CTGAN",
    "fraud detection",
    "privacy-preserving machine learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]
dependencies = [
    "torch>=2.0.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "scipy>=1.11.0,<2.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "ctgan==0.12.0",
    "opacus==1.5.4",
    "matplotlib>=3.7.0,<4.0.0",
    "seaborn>=0.12.0,<1.0.0",
    "PyYAML>=6.0",
    "pyarrow>=13.0.0",
    "statsmodels>=0.14.0",
    "tqdm>=4.65.0",
]

[project.urls]
Repository = "https://github.com/<your-user>/privacy-preserving-federated-fraud-detection"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
